Fix standard DKT

+++ Knowledge_Tracing/code/models/DKT_models/DKT/standard/data_utils.py
@@ -9,7 +9,7 @@ from Knowledge_Tracing.code.data_processing.load_preprocessed.load_preprocessed_
 
 
 def create_dataset(generator, features_depth, skill_depth, shuffle=True, batch_size=1024):
-    input_types = {"feature": tf.float32}
+    input_types = {"feature": tf.int32}
     output_types = {"target_label": tf.float32, "target_skill": tf.float32}
 
     input_shapes = {"feature": [None, features_depth], }
